refactor: CMD-189 move utils and README to root
