+++ taccsite_cms/management/commands/group_perms/grid_editor_basic.py
@@ -4,7 +4,7 @@ To edit layout (Containers, Rows, Columns)
 
 from django.contrib.auth.models import Group
 
-from .util import (
+from ..util import (
     let_view_and_change_page_structure,
     let_view_and_change_grid
 )
