+++ taccsite_cms/management/commands/group_perms/media_editor_advanced.py
@@ -5,7 +5,7 @@ To add/delete images, videos, thumbnail sizes, and folders
 
 from django.contrib.auth.models import Group
 
-from .util import (
+from ..util import (
     let_view_and_change_page_structure,
     let_view_and_change_media_plugins,
     let_add_and_delete_media_plugins,
