+++ taccsite_cms/management/commands/group_perms/media_editor_basic.py
@@ -4,7 +4,7 @@ To edit images, videos, thumbnail sizes, and folders
 
 from django.contrib.auth.models import Group
 
-from .util import (
+from ..util import (
     let_view_and_change_page_structure,
     let_view_and_change_media_plugins,
     let_view_and_change_adv_media_plugins,
