+++ taccsite_cms/management/commands/group_perms/news_writer_advanced.py
@@ -6,7 +6,7 @@ To add/delete news Categories and Tags
 
 from django.contrib.auth.models import Group
 
-from .util import (
+from ..util import (
     add_perm,
     let_view_and_change_page_structure,
     let_view_and_change_plugin,
