+++ taccsite_cms/management/commands/group_perms/text_editor_basic.py
@@ -4,7 +4,7 @@ To edit textual elements blocks, folders, and files
 
 from django.contrib.auth.models import Group
 
-from .util import (
+from ..util import (
     let_view_and_change_page_structure,
     let_view_and_change_text,
     let_view_and_change_folder,
